@@ -77,6 +77,8 @@
   "author": "David Zukowski <david@zuko.me> (http://zuko.me)",
   "license": "MIT",
   "dependencies": {
+    "@terebentina/react-popover": "^1.3.0",
+    "axios": "^0.13.1",
     "babel-cli": "^6.5.1",
     "babel-core": "^6.3.17",
     "babel-loader": "^6.2.0",
@@ -87,6 +89,10 @@
     "babel-preset-stage-0": "^6.3.13",
     "babel-register": "^6.3.13",
     "babel-runtime": "^6.3.19",
+<<<<<<< HEAD
+=======
+    "classnames": "^2.2.5",
+>>>>>>> with-react-toolbox
     "better-npm-run": "0.0.10",
     "css-loader": "^0.23.0",
     "cssnano": "^3.3.2",
@@ -96,35 +102,57 @@
     "fs-extra": "^0.30.0",
     "history": "^2.0.0",
     "html-webpack-plugin": "^2.7.1",
+    "immutable": "^3.8.1",
     "imports-loader": "^0.6.5",
     "ip": "^1.1.2",
+    "isomorphic-fetch": "^2.2.1",
     "json-loader": "^0.5.4",
     "koa": "^2.0.0-alpha.3",
     "koa-connect-history-api-fallback": "^0.3.0",
     "koa-convert": "^1.2.0",
     "koa-proxy": "^0.6.0",
+<<<<<<< HEAD
     "koa-static": "^3.0.0",
+=======
+    "koa-router": "^7.0.0",
+    "koa-static": "^3.0.0",
+    "material-ui": "^0.15.0",
+    "moment": "^2.13.0",
+>>>>>>> with-react-toolbox
     "node-sass": "^3.7.0",
     "normalize.css": "^4.1.1",
     "postcss-loader": "^0.9.0",
-    "react": "^15.0.0",
+    "purecss": "0.6.0",
+    "react": "^15.1.0",
+    "react-addons-css-transition-group": "^15.0.1",
+    "react-autosuggest": "^3.7.3",
+    "react-cardstack": "^0.1.1",
+    "react-datetime": "^2.3.1",
     "react-dom": "^15.0.0",
+    "react-facebook-login": "^3.1.0",
+    "react-modal": "^1.4.0",
+    "react-motion": "^0.4.4",
     "react-redux": "^4.0.0",
     "react-router": "^2.2.0",
     "react-router-redux": "^4.0.0",
+    "react-spinkit": "^1.1.8",
+    "react-star-rating-component": "^1.0.0",
+    "react-tap-event-plugin": "^1.0.0",
+    "react-toggle": "^2.1.0",
+    "react-toolbox": "^1.0.2",
+    "redbox-react": "^1.2.10",
     "redux": "^3.0.0",
     "redux-thunk": "^2.0.0",
     "rimraf": "^2.5.1",
     "sass-loader": "^4.0.0",
+<<<<<<< HEAD
+=======
+    "socket.io-client": "^1.4.8",
+>>>>>>> with-react-toolbox
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.6",
     "webpack": "^1.12.14",
-    "yargs": "^4.0.0",
-    "immutable": "^3.8.1",
-    "react-autosuggest": "^3.7.3",
-    "koa-router": "^7.0.0",
-    "purecss": "0.6.0",
-    "materialize-css":"0.97.6"
+    "yargs": "^4.0.0"
   },
   "devDependencies": {
     "babel-eslint": "^6.0.0-beta.6",
@@ -148,11 +176,22 @@
     "karma-mocha-reporter": "^2.0.0",
     "karma-phantomjs-launcher": "^1.0.0",
     "karma-webpack-with-fast-source-maps": "^1.9.2",
+<<<<<<< HEAD
+=======
+    "less": "^2.7.1",
+    "less-loader": "^2.2.3",
+>>>>>>> with-react-toolbox
     "mocha": "^3.0.1",
     "nodemon": "^1.8.1",
     "phantomjs-prebuilt": "^2.1.3",
     "react-addons-test-utils": "^15.0.0",
+<<<<<<< HEAD
     "redbox-react": "^1.2.10",
+=======
+    "redux-devtools": "^3.3.1",
+    "redux-devtools-dock-monitor": "^1.1.1",
+    "redux-devtools-log-monitor": "^1.0.11",
+>>>>>>> with-react-toolbox
     "sinon": "^1.17.3",
     "sinon-chai": "^2.8.0",
     "webpack-dev-middleware": "^1.6.1",
