@@ -3,46 +3,116 @@ import { IndexLink, Link } from 'react-router'
 import FileIcons from '../assets/file-icons.png'
 import Autosuggest from 'react-autosuggest'
 import AsyncAutocomplete from 'components/AsyncAutocomplete'
+import store from 'store/createStore'
+import IconButton from 'material-ui/IconButton'
+import CommunicationLocationOn from 'material-ui/svg-icons/communication/location-on'
+import Spinner from 'react-spinkit'
+import classes from './HomeView.scss'
+import classNames from 'classnames';
 
 const HomeView = React.createClass({
     getInitialState() {
         return {
-                userSearchChange: this.props.userZipSearchChange,
-                apiUrl: "/api/city/",
-                searchTextAlreadyInStore:this.props.userZipSearch.get('searchText')
+                fetchingGeoZips:false,
         };
     },
+    contextTypes: {
+    	router: React.PropTypes.object.isRequired
+  	},
+    componentDidMount() {
+       
+    },
+    
+    goToProviderProfile(){
+    	this.context.router.push('/provider');
+    },
+    getLocationCordinates(){
+    	let self = this;
+    	this.setState({
+    		fetchingGeoZips:true
+    	})
+        if (navigator.geolocation) {
+            navigator.geolocation.getCurrentPosition(function(pos) {     
+            let latitude =  pos.coords.latitude,
+            	longitude =  pos.coords.longitude;
+            fetch('https://nominatim.openstreetmap.org/reverse?format=json&lat='+latitude+'&lon='+longitude+'&addressdetails=1')
+            	.then(function(data){
+            		return data.json();
+            	}).then(function(response){
+	            	self.props.userZipSearchChange(response.address.postcode);
+	            	self.setState({
+			    		fetchingGeoZips:false
+			    	})
+            	})
+        	},
+	        function(err) {
+	            this.setState({
+    				fetchingGeoZips:true
+    			})
+	        });
+        } else {
+            // show some error message
+        }
+    },
     render() {
+    	console.log("parent rerender",this.state);
         return (
             <div>
-				<div className="splash-container counter-fixed-menu">
-					<div className="banner-wrapper">
+				<div className={classNames(classes["splash-container"],classes["pure-override-letter-spacing"])}>
+					<div className={classes["banner-wrapper"]}>
 						<div className="pure-g">
 							<div className = "pure-u-1 pure-u-md-1-2">
-							    <div className="splash ">
-							        <h1 className="splash-head">
+							    <div className={classes["splash"]}>
+							        <h1 className={classes["splash-head"]}>
 							        	Please enter a zip code to find food close to you.
 							        </h1>
-							        <AsyncAutocomplete settings={this.state}/>
-							        <p className="is-center">
+							        <AsyncAutocomplete settings={{
+							        	userSearchText : this.props.homepage.get('userZipSearch'),
+							        	apiUrl:'/api/locations/zipcodeTypeAssist',
+							        	action:this.props.userZipSearchChange
+							        }}/>
+							        <IconButton
+							        	style = {{
+							        			padding:'0px',
+							        			height:'0px',
+							        			width:'0px',
+							        			marginLeft:'-30px',
+							        			top:'6px',
+							        			display:'inline-block',
+							        			visibility:(this.state.fetchingGeoZips)? 'hidden':'initial'							        		
+							        		}}
+							        	onClick = {this.getLocationCordinates}
+							        >
+							        	<CommunicationLocationOn/>
+							        </IconButton>
+							        <Spinner spinnerName='circle' 
+							        	style = {{	display:'inline-block',
+							        				visibility:(this.state.fetchingGeoZips)?'initial':'hidden',
+							        				top:'5px'
+							        			}}
+							        />
+							        <p className={classes["is-center"]}>
 							        	<button className="pure-button pure-button-primary">Get Started</button>
 							        </p>
 							    </div>
 						    </div>
 						    <div className = "pure-u-1 pure-u-md-1-2">
-							    <div className="splash">
-							        <div className="splash-head">
+							   <div className={classes["splash"]}>
+							        <div className={classes["splash-head"]}>
 							        	Advertise your food right away in 3 easy steps
 							        </div>
-							        <p className="is-center">
-							        	<button className="pure-button pure-button-primary">Get Started</button>
+							        <p className={classes["is-center"]}>
+							        	<button className="pure-button pure-button-primary"
+							        		onClick={this.goToProviderProfile}>
+							        		Get Started
+							        	</button>
 							        </p>
 							    </div>
 						    </div>
 						</div>
 					</div>
 				</div>
-				<div className="content-wrapper">
+				{/*<div className="content-wrapper">
 				    <div className="content">
 				        <h2 className="content-head is-center">Excepteur sint occaecat cupidatat.</h2>
 
@@ -150,17 +220,16 @@ const HomeView = React.createClass({
 				        View the source of this layout to learn more. Made with love by the YUI Team.
 				    </div>
 
-				</div>
+				</div>*/}
 			</div>
         )
     }
 })
 
 HomeView.propTypes = {
-    leftNavOpen: React.PropTypes.bool.isRequired,
-    leftnavstatechange: React.PropTypes.func.isRequired,
-    userZipSearchChange: React.PropTypes.func.isRequired,
-    userZipSearch: React.PropTypes.object.isRequired
+	globalState:React.PropTypes.object.isRequired,
+   	homepage:React.PropTypes.object.isRequired,
+   	userZipSearchChange:React.PropTypes.func.isRequired
 }
 
 export default HomeView
