@@ -1,13 +1,16 @@
 import React, { Component, PropTypes } from 'react';
 import classes from './providerentryform.scss';
 import { email, maxLength, required } from './../../utils/formValidation';
+import Toggle from 'react-toggle';
+import classNames from 'classnames';
+import Modal from 'react-modal';
+import {RadioButton, RadioButtonGroup} from 'material-ui/RadioButton';
+import axios from 'axios';
 const maxCount = 100;
-
-
 class ProviderEntryForm extends React.Component {
     constructor(props) {
-        super(props); 
-        this.state= this.props.providerEntryForm;
+        super(props);
+        this.state = this.props.providerEntryForm;
         this.mapFieldsToValidationType = {
             title: required,
             emailId: email,
@@ -17,144 +20,329 @@ class ProviderEntryForm extends React.Component {
         this.handleChange = this.handleChange.bind(this);
         this.handleFocus = this.handleFocus.bind(this);
         this.formSubmit = this.formSubmit.bind(this);
-        this.changeTitle = this.changeTitle.bind(this);
+        this.changeStoreVal = this.changeStoreVal.bind(this);
+        this.toggle = this.toggle.bind(this);
+        this.validateForm = this.validateForm.bind(this);
     }
-
     handleChange(event) {
         let input = event.target.value;
         let stateKeyName = event.target.name;
+        let actionObj = {
+            storeKey :stateKeyName,
+            payload : input
+        }
         let validation = this.mapFieldsToValidationType[stateKeyName];
-        let errorMsg = validation(input);
+        let errorMsg;
+        if (validation) {
+            errorMsg = validation(input);
+        }
         if (errorMsg) {
             let errorMsgkey = stateKeyName + 'ErrorMsg';
-            this.setState({
-                [errorMsgkey]: errorMsg
+            this.props.addProviderErrorMsg({
+                storeKey:errorMsgkey,
+                payload:errorMsg
             });
-            this.setState({
-                stateKeyName: input
-            });
-            this.props.providerEntryForm[stateKeyName] = input;
-            this.props.providerEntryForm[errorMsgkey] = errorMsg  // hacky again
-        } else {
-            this.setState({
-                [stateKeyName]: input
-            })
-            this.props.providerEntryForm[stateKeyName] = input;
         }
+        this.props.addProviderInfo({
+            storeKey:stateKeyName,
+            payload:input
+        })
+       
     }
-
+    toggle(storeKey) {
+        this.props.addProviderInfo({
+            storeKey: storeKey,
+            payload:!this.props.providerEntryForm.get(storeKey)
+        })
+    }
     handleFocus(event) {
         let stateKeyName = event.target.name;
         if (stateKeyName) {
             // clear the error msg if it exists
             let errorMsgkey = stateKeyName + 'ErrorMsg';
-            this.setState({
-                [errorMsgkey]: null
+            this.props.addProviderErrorMsg({
+                storeKey:errorMsgkey,
+                payload:null
             });
-            this.props.providerEntryForm[errorMsgkey] = null;
         }
     }
-
     setCount(event) {
         let input = event.target.value;
         let stateKeyName = event.target.name;
         this.setState({
-        	description:input,
+            description: input,
             chars_left: maxCount - input.length
         })
-        this.props.providerEntryForm.chars_left = maxCount - input.length;
-        this.props.providerEntryForm.description = input;
+        this.props.providerEntryForm.chars_left = maxCount - input.length; // hacky 
+        this.props.providerEntryForm.description = input; // hacky 
     }
 
-    changeTitle(event){
-    	let input = event.target.value;
-    	this.setState({
-    		title:input
-    	});
-    	// THIS IS REALLY HACKY .. 
-    	// IDEALLY WE SHOULD HAVE SHARED A REFERENCE TO REDUCER WHICH MAKES CHANGES IN THE STORE
-    	// now we are being forced to keep state internally and globally
-    	this.props.providerEntryForm.title=input;
+    changeStoreVal(event) {
+        let input = event.target.value;
+        let stateKeyName = event.target.name;
+        this.props.addProviderInfo({
+            storeKey:stateKeyName,
+            payload:input
+        });
     }
-
-    formSubmit(event) {
+    validateForm(){
         let self = this;
         let noErrorsInform = true;
         for (let key in this.mapFieldsToValidationType) {
             if (this.mapFieldsToValidationType.hasOwnProperty(key)) {
-                let errorMsg = this.mapFieldsToValidationType[key](this.state[key]);
-                console.log(errorMsg + '  for key: '+ key +'  whose value is '+ this.state[key]);
+                let errorMsg = this.mapFieldsToValidationType[key](this.props.providerEntryForm.get(key));
                 if (errorMsg) {
                     noErrorsInform = false;
-                    let errorStateKey = [key] + 'ErrorMsg'
-                    self.setState({
-                        [errorStateKey]: errorMsg
-                    });
-                    this.props.providerEntryForm[errorStateKey]=errorMsg;
+                    let errorStateKey = [key] + 'ErrorMsg';
+                    self.props.addProviderErrorMsg({
+                        storeKey:errorStateKey,
+                        payload:errorMsg
+                    })
                 }
             }
         }
-        if (!noErrorsInform) {
-            this.setState({
-                allClear: false
-            })
-            this.props.providerEntryForm.allClear = false;
-        } else {
-            this.setState({
-                allClear: true
-            })
-            this.props.providerEntryForm.allClear = true;
+        return noErrorsInform;
+    }
+    formSubmit(event) {
+        if(this.validateForm()){
+            let token = sessionStorage.getItem('token');
+            axios({
+              method: 'post',
+              headers: {
+                      'Content-Type': 'application/json',
+                      'Authorization': 'Bearer ' + token
+                  },
+              url:'/api/providers/registration',
+              data: this.props.providerEntryForm.toJS()
+            }); 
+            this.props.addProviderEntryState({
+                storeKey: "stepIndex",
+                payload:1
+            });
         }
     }
     render() {
-        let { chars_left, title, description, streetName, crosStreetName, city, emailId, titleErrorMsg, descriptionErrorMsg, cityErrorMsg, emailIdErrorMsg } = this.state;
-        console.log(title);
+        let { chars_left, title, description, streetName, crosStreetName, city, emailId, titleErrorMsg, descriptionErrorMsg, cityErrorMsg, emailIdErrorMsg, keepEmailPrivateFlag, keepAddressPrivateFlag,pickUpFlag,pickUpAddtnlComments, includeAddressInEmail, deliveryAddtnlComments,deliveryMinOrder,deliveryRadius,allClear,providerAddressJustificationModalOpen,doYouDeliverFlag } = this.props.providerEntryForm.toJS();
+        const styles = {
+          block: {
+            maxWidth: 250,
+          },
+          radioButton: {
+            marginBottom: 12,
+          },
+        };
         return (
-
             <div>
-            <form className="pure-form">
-	         	<fieldset className="pure-group">
-			        <input type="text"  className="pure-u-1" placeholder="title (required)" name="title" value={title}
-			        onChange={this.changeTitle}
-			        onBlur={this.handleChange} 
-			        onFocus={this.handleFocus}
-			    />
-			        <span className = {classes["error-message"]}>{(titleErrorMsg)?'*'+titleErrorMsg:undefined}</span>
-			        <textarea className = "pure-u-1"name="description" placeholder="Background (optional)" value={description}
-			        	onBlur={this.handleChange} 
-			        	onFocus={this.handleFocus} 
-			        	onChange={this.setCount.bind(this)} 
-			        >	
-			        </textarea>
-			        <span className = {classes["error-message"]}>{(descriptionErrorMsg)?'*'+descriptionErrorMsg:undefined}</span>
-			        <div>{chars_left}/100</div>
-			    </fieldset>
-			    <fieldset className="pure-group">
-			    	<legend className={classes["pull-left"]}>
-			    		Display your neighbouring address
-			    	</legend>
-			        <input type="text"  style = {{marginBottom:0.5+'em'}} name="streetName"  placeholder="Street Name (optional)" />
-			        <input type="text"  style = {{marginBottom:0.5+'em'}} name="crosStreetName" placeholder="Cross Street Name (optional)"/>
-			        <input type="text"  name="city"  placeholder="City (required)" style = {{marginBottom:0.5+'em'}}
-			        	onBlur={this.handleChange} 
-			        	onFocus={this.handleFocus}/>
-			        <span className = {classes["error-message"]}>{(cityErrorMsg)?'*'+cityErrorMsg:undefined}</span>
-			    </fieldset>
-			    <fieldset className="pure-group">
-			    	<legend className={classes["pull-left"]}>
-			    		Keep my email private
-			    	</legend>
-			        <input type="text" name="emailId" placeholder="email (required)" style = {{marginBottom:0.5+'em'}} 
-			        	onBlur={this.handleChange} 
-			        	onFocus={this.handleFocus}/>
-			        <span className = {classes["error-message"]}>{(emailIdErrorMsg)?'*'+emailIdErrorMsg:undefined}</span>
-			    </fieldset>
-			</form>
-		</div>
+                <form className="pure-form pure-form-stacked">
+                    <fieldset className="pure-group">
+                        <input type="text"  className="pure-u-1" placeholder="*title" name="title" value={this.props.providerEntryForm.get('title')}
+                        onChange={this.changeStoreVal}
+                        onBlur={this.handleChange} 
+                        onFocus={this.handleFocus}
+                    />
+                        <span className = {classes["error-message"]}>{(titleErrorMsg)?'*'+titleErrorMsg:undefined}</span>
+                        <textarea className = "pure-u-1"name="description" placeholder="background" value={description}
+                            onBlur={this.handleChange} 
+                            onFocus={this.handleFocus} 
+                            onChange={this.changeStoreVal} 
+                        >   
+                        </textarea>
+                        <span className = {classes["error-message"]}>{(descriptionErrorMsg)?'*'+descriptionErrorMsg:undefined}</span>
+                        <div>{chars_left}/100</div>
+                    </fieldset>
+                    <fieldset className="pure-group">
+                        <input type="text" name="emailId" placeholder="*email" style = {{marginBottom:0.5+'em'}} value={emailId}
+                            onBlur={this.handleChange} 
+                            onFocus={this.handleFocus}
+                            onChange={this.changeStoreVal}
+                            className={"pure-u-1"}/>
+                        <span className = {classes["error-message"]}>{(emailIdErrorMsg)?'*'+emailIdErrorMsg:undefined}</span>
+                    </fieldset>
+                    <fieldset className="pure-group">
+                        <legend className={classes["pull-left"]}>
+                            <div>
+                                Address:
+                            </div>
+                            <div>
+                                <span style={{fontSize:'0.75em'}}>Display this on my public profile page</span>
+                                <Toggle
+                                defaultChecked={!keepAddressPrivateFlag}
+                                onChange={()=>{this.toggle('keepAddressPrivateFlag')}}
+                                className = {classes["input-hidden"]} 
+                                />
+                                <a className={classNames("pure-menu-link",classes["address-justification"])} 
+                                    onClick={()=>{this.toggle('providerAddressJustificationModalOpen')}}>
+                                    why we need it?
+                                </a> 
+                            </div>
+                            
+                        </legend>
+                        <input type="text"  name="streetName" placeholder="address line1" value = {streetName}
+                            onChange={this.changeStoreVal}
+                            className="pure-u-1" 
+                         />
+                        <input type="text" name="crosStreetName" placeholder="address line 2" value = {crosStreetName}
+                            onChange={this.changeStoreVal}
+                            className="pure-u-1" 
+                        />
+                        <input type="text"  name="city"  placeholder="*city" value = {city}
+                            onBlur={this.handleChange} 
+                            onFocus={this.handleFocus}
+                            onChange={this.changeStoreVal}
+                            className="pure-u-1"
+                        />
+                        <span className = {classes["error-message"]}>{(cityErrorMsg)?'*'+cityErrorMsg:undefined}</span>
+                    </fieldset>
+                    {(keepAddressPrivateFlag)?
+                        <fieldset className="pure-group">
+                            <div>
+                                <p>
+                                    Please note that the address entered above will be used as a pick-up address
+                                </p>
+                                <p>
+                                    Choose from one of the following privacy options
+                                </p>
+                                <RadioButtonGroup name="shipSpeed" 
+                                    defaultSelected={includeAddressInEmail.toString()}
+                                    onChange={(event)=>this.toggle('includeAddressInEmail')}
+                                >
+                                  <RadioButton
+                                    value="true"
+                                    label="don't display the address but include it in the email sent to customer after order submission"
+                                    style={styles.radioButton}
+                                  />
+                                  <RadioButton
+                                    value="false"
+                                    label="i will cordinate with customer regarding the pick-up location"
+                                    style={styles.radioButton}
+                                    
+                                  />
+                                </RadioButtonGroup>
+                            </div>      
+                        </fieldset>
+                        : undefined
+                    }
+                    <fieldset className="pure-group">
+                        We are pick-up service
+                        <Toggle
+                            defaultChecked={pickUpFlag}
+                            onChange={()=>{this.toggle('pickUpFlag')}}
+                            className = {classes["input-hidden"]}
+                        />
+                    </fieldset>
+                    {(pickUpFlag)?
+                        <div>
+                            <fieldset className = "pure-group">
+                                <textarea className = "pure-u-1" name="pickUpAddtnlComments" 
+                                    placeholder="additional comments about pick-up" value={pickUpAddtnlComments}
+                                    onBlur={this.handleChange} 
+                                    onFocus={this.handleFocus} 
+                                    onChange={this.changeStoreVal} 
+                                >
+                                </textarea>
+                            </fieldset>
+                        </div>
+                        :
+                        undefined
+                    }
+                    <fieldset className="pure-group">
+                        <legend className={classes["pull-left"]}>
+                                We can deliver 
+                            <Toggle
+                                defaultChecked={doYouDeliverFlag}
+                                onChange={()=>{this.toggle('doYouDeliverFlag')}}
+                                className = {classes["input-hidden"]} 
+                            /> 
+                        </legend>
+                            {(doYouDeliverFlag)?
+                                <div>
+                                    <div>
+                                        <div className="pure-u-1 pure-u-md-1-2">
+                                            <label>
+                                                Delivery within
+                                            </label>
+                                            <div>
+                                                <select id="delivery"  className="pure-u-3-4"
+                                                    name="deliveryRadius"
+                                                    onBlur={this.handleChange} 
+                                                    onFocus={this.handleFocus}
+                                                    onChange={this.changeStoreVal}
+                                                    value={deliveryRadius}
+                                                >
+                                                    <option  value="0">Please select miles</option>
+                                                    <option>5</option>
+                                                    <option>10 </option>
+                                                    <option>15</option>
+                                                    <option>20</option>
+                                                    <option>25</option>
+                                                    <option>30</option>
+                                                </select>
+                                            </div>
+                                        </div>
+                                        <div className="pure-u-1 pure-u-md-1-2">
+                                            <label>Minimum Order</label>
+                                            <input id="min-order" placeholder="Example: 25"className="pure-u-3-4" type="text"
+                                                name="deliveryMinOrder"
+                                                onBlur={this.handleChange} 
+                                                onFocus={this.handleFocus}
+                                                onChange={this.changeStoreVal}
+                                                value={deliveryMinOrder}
+                                            />
+                                        </div>
+                                    </div>   
+                                    <textarea className = "pure-u-1" style={{marginTop:'10px'}} 
+                                        name="deliveryAddtnlComments" 
+                                        placeholder="Please add comments like delivery charges. 
+                                        Example:
+                                        Delivery Fee $6 
+                                        Free above $100"
+                                        value={deliveryAddtnlComments}
+                                        style = {{marginTop:'10px'}}
+                                        onBlur={this.handleChange} 
+                                        onFocus={this.handleFocus} 
+                                        onChange={this.changeStoreVal} 
+                                    >
+                                    </textarea>
+                                </div> 
+                            :
+                                undefined
+                        }                            
+                    </fieldset>
+                </form>
+                <Modal
+                  isOpen={providerAddressJustificationModalOpen}
+                  onRequestClose={()=>{this.toggle('providerAddressJustificationModalOpen')}}
+                >
+                  <div ref="subtitle"
+                    style={{
+                      marginBottom:'10%'
+                    }}
+                  >
+                    <p>
+                        Address is used to match you with:
+                        <ul>
+                            <li>
+                                customers searching for food in the area. 
+                            </li>
+                        </ul> 
+                    </p>
+                    <p>
+                         Still not convinced. Please feel free to give ur your cross street address.
+                         <br/>
+                         However make sure to cordinate with your customer regarding the accurate location for pick-up
+                    </p>
+                  <p>
+                        fillurtummy is dedicated to keeping your data private
+                  </p>
+
+                  </div>
+                </Modal>
+            </div>
         )
     }
 }
 ProviderEntryForm.propTypes = {
     providerEntryState: React.PropTypes.object.isRequired,
+    providerEntryForm: React.PropTypes.object.isRequired
 };
 export default ProviderEntryForm;
