@@ -1,18 +1,13 @@
 // ------------------------------------
 // Constants
 // ------------------------------------
-export const LEFTNAV_OPEN_CLOSE = 'LEFTNAV_OPEN_CLOSE';
 export const USER_ZIP_SEARCH_CHANGE = 'USER_ZIP_SEARCH_CHANGE'
 
 import { Map, fromJS } from 'immutable'
 // ------------------------------------
 // Actions
 // ------------------------------------
-export function leftnavstatechange(state) {
-    return {
-        type: LEFTNAV_OPEN_CLOSE
-    }
-}
+
 export function userZipSearchChange(val) {
     return {
         type: USER_ZIP_SEARCH_CHANGE,
@@ -21,7 +16,6 @@ export function userZipSearchChange(val) {
 }
 
 export const actions = {
-    leftnavstatechange,
     userZipSearchChange
 }
 
@@ -29,9 +23,6 @@ export const actions = {
 // Action Handlers
 // ------------------------------------
 const ACTION_HANDLERS = {
-    [LEFTNAV_OPEN_CLOSE]: (state, action) => {
-        return state.set('leftNavOpen', !state.get('leftNavOpen'));
-    },
     [USER_ZIP_SEARCH_CHANGE]:(state,action)=>{
         let newState = state.updateIn(['userZipSearch','searchText'],value=>action.payload);
         return newState;
@@ -42,7 +33,6 @@ const ACTION_HANDLERS = {
 // Reducer
 // ------------------------------------
 let initialState = Map({
-    leftNavOpen: false,
     userZipSearch: Map({
         searchText: ''
     })
