@@ -1,9 +1,9 @@
 // We only need to import the modules necessary for initial render
 import CoreLayout from '../layouts/CoreLayout/CoreLayout'
 import Home from './Home'
-import CounterRoute from './Counter'
+import ChatRoute from './Chat'
 import ProviderRoute from './Provider'
-
+import ProviderProfilePage from './ProviderProfilePage'
 
 /*  Note: Instead of using JSX, we recommend using react-router
     PlainRoute objects to build route definitions.   */
@@ -13,8 +13,9 @@ export const createRoutes = (store) => ({
   component: CoreLayout,
   indexRoute: Home,
   childRoutes: [
-    CounterRoute(store),
-    ProviderRoute(store)
+    ChatRoute(store),
+    ProviderRoute(store),
+    ProviderProfilePage(store)
   ]
 })
 
